requirements.txt
apache-airflow==2.8.1
dbt-postgres==1.7.4
psycopg2-binary==2.9.9
requests==2.31.0
tenacity==8.2.3
pandas==2.2.0
python-dotenv==1.0.1
